feat: add CF domain creation
There are a couple of missing things, which I detailed in some TODOS
you can find through the code. When it comes to the changes, it works
fine for CloudFlare, but we could do as its done with DO, and retrieve
data beforehand, in order to catch bugs before doing any modifications.